requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-phonenumber-field==7.0.1
docopt==0.6.2
idna==3.4
phonenumbers==8.13.2
pipreqs==0.4.11
pycodestyle==2.10.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
yarg==0.1.9
